Fix #14 - add support for logging to Logger32
The default option has been replaced with a TCP options. The ADIF data
sent is the same but the socket type is now TCP as opposed to UDP. The
Log4om is renamed to UDP.

Also fix an order bug with upgrading the DB via alembic in the db init code.
If the _config_ table was updated it would error b/c init_config was called
before init_alembic_ver