Support for light and dark mode toggle
Had to move react elements into a seperate file / react.ts function so we
can switch themes on the fly. Also there's a bug with AppBar where the
button colors for light mode are set to the same as bg color of Appbar
making the buttons disappear.

Fixes #31
Also fix deprecation warning in index.jsx and other ts warnings