Fix for startup issues on Linux
The pywebviewready JS event does not fire in the same way that it does on windows
It seems that the event doesnt fire for the child react components on Linux.
Linux it needs to call direct. but it messes up when the refresh button is clicked
so more checks need to be done to keep from crashing.