OP STATS fixes. Update adif-io py lib to latest
The adif import of OP was kinda busted. It's dumb to handle anything
that doesn't have a SIG set. Before I had logs that had a park in
SIG_INFO but no SIG. Fix the log instead of trying to parse tens of thousands
of qsos. This filters it down a lot and also handles errors better.
Fixes #78