Workflow fixes (#82)
* Add installation of libgirepository-2.0-dev

* Remove installation of libgirepository-2.0-dev

This is not available on ubuntu-22.04

* Add GitHub Actions workflow for Node.js on Ubuntu 24

This workflow installs node dependencies, builds the source code on Ubuntu 24.04.

* Update ubuntu-build.yml

Change ubuntu image to 24.04

* Remove default init:gtk command to install now failing libs

remove pycairo and pygobject from original init:gtk npm script.
This was causing the CI to pull in new libs that seem to be breaking b/c of new depedency problems. Requirements.txt already pinned this pip libs to a good version.