Update alerts to add callsign search and band exclusions
Now users can add a callsign to alerts as well as well as limit location
alerts to more likely bands. People in the same state probably dont care
for alerts for new parks in their states unless the activator is on lower
band. Also fix a un-reported bug w/ wwbota regions in dropdown filter

Fix #87 and #65