Updates for ubuntu 24.04 and linux builds
The macos build updates broke the 24.04 ubuntu build and the fix was to update
pyinstaller to latests. Tested on the three platforms that use pyinstaller and
it seems to work ok.

This also fixes #110 (or at least appears to have fixed it in my VM). Presumably
the newer pyinstaller pulled in the needed TLS modules where the older version didnt.

Also fix data in dist for linux builds.